requirements.txt
# 市场自动化工具 - Python依赖包

# 图像处理和计算机视觉
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0

# OCR文字识别
pytesseract>=0.3.8
# 可选OCR引擎
# paddlepaddle>=2.2.0
# paddleocr>=2.5.0
# easyocr>=1.6.0

# ADB控制和设备交互
# 注意：ADB工具需要单独安装到系统PATH中
# pure-python-adb>=0.2.0  # 可选，纯Python ADB实现
uiautomator2>=2.16.0  # Android UI自动化框架

# 旧架构依赖（可选，用于回滚）
# frida-tools>=12.0.0  # 已废弃，仅用于兼容性

# 系统交互
psutil>=5.8.0

# 日期时间处理
python-dateutil>=2.8.0

# JSON处理增强
ujson>=5.0.0

# 日志处理
colorlog>=6.0.0

# 配置文件处理
pyyaml>=6.0

# 网络请求
requests>=2.25.0

# 数据处理
pandas>=1.3.0

# 数据库 (sqlite3是Python标准库，无需安装)
# sqlite3

# 并发处理 (concurrent-futures是Python标准库，无需安装)
# concurrent-futures>=3.1.1

# 文件监控
watchdog>=2.1.0

# 进程管理
pywin32>=304; sys_platform == "win32"

# 图像处理增强
scikit-image>=0.19.0
# 图像特征匹配和模板匹配
# imutils>=0.5.4

# 数据可视化（用于调试和分析）
matplotlib>=3.5.0
seaborn>=0.11.0

# 性能分析和优化
memory-profiler>=0.60.0
py-spy>=0.3.0

# 测试框架（开发依赖）
pytest>=6.2.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# 代码质量工具（开发依赖）
flake8>=4.0.0
black>=22.0.0
isort>=5.10.0

# 类型检查（开发依赖）
mypy>=0.950

# 文档生成（开发依赖）
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0